privileges added

+++ routes/projectRoute.js
@@ -34,45 +34,50 @@ router.get("/:id", (req, res) => {
 
 //ADD A PROJECT
 router.post("/", middleware, (req, res) => {
-  try {
-    let sql = `INSERT INTO projects (title, description, type, tech, email, postedBy) VALUES(? ,? , ? , ? , ?, ? );`;
-    let { title, description, type, tech, email, postedBy } = req.body;
-    // const date = new Date().toISOString().slice(0, 10);
-    let project = {
-      title: title,
-      description: description,
-      type: type,
-      tech: tech,
-      email: email,
-      postedBy: postedBy,
-    };
-    con.query(
-      sql,
-      [
-        project.title,
-        project.description,
-        project.type,
-        project.tech,
-        project.email,
-        project.postedBy,
-      ],
-      (err, result) => {
-        if (err) throw err;
-        console.log(result);
+  if (req.user.userRole === "client" || "admin") {
+    try {
+      let sql = `INSERT INTO projects (title, description, type, tech, email, postedBy) VALUES(? ,? , ? , ? , ?, ? );`;
+      let { title, description, type, tech, email, postedBy } = req.body;
+      // const date = new Date().toISOString().slice(0, 10);
+      let project = {
+        title: title,
+        description: description,
+        type: type,
+        tech: tech,
+        email: email,
+        postedBy: postedBy,
+      };
+      con.query(
+        sql,
+        [
+          project.title,
+          project.description,
+          project.type,
+          project.tech,
+          project.email,
+          project.postedBy,
+        ],
+        (err, result) => {
+          if (err) throw err;
+          console.log(result);
 
-        res.json({
-          msg: "Project added Successfully",
-        });
-      }
-    );
-  } catch (error) {
-    console.log(error);
+          res.json({
+            msg: "Project added Successfully",
+          });
+        }
+      );
+    } catch (error) {
+      console.log(error);
+    }
+  } else {
+    res.send("not authorised");
   }
 });
 
 // UPDATE A PROJECT
 router.patch("/:id", middleware, (req, res) => {
-  if (userRole === client) {
+  if (req.user.userRole === "client" || "admin") {
+    // if (userRole === "client") {
     try {
       let sql = "SELECT * FROM projects WHERE ? ";
       let project = { project_id: req.params.id };
@@ -87,7 +92,7 @@ router.patch("/:id", middleware, (req, res) => {
             description: req.body.description,
             type: req.body.type,
             tech: req.body.tech,
-            email:req.body.email,
+            email: req.body.email,
             postedBy: req.body.postedBy,
           };
           con.query(updateSql, updateProject, (err, updated) => {
@@ -109,15 +114,19 @@ router.patch("/:id", middleware, (req, res) => {
 
 // DELETE PROJECT
 router.delete("/:id", middleware, (req, res) => {
-  try {
-    let sql = "Delete from projects WHERE ?";
-    let project = { project_id: req.params.id };
-    con.query(sql, project, (err, result) => {
-      if (err) throw err;
-      res.send(result);
-    });
-  } catch (error) {
-    console.log(error);
+  if (req.user.userRole === "client" || "admin") {
+    try {
+      let sql = "Delete from projects WHERE ?";
+      let project = { project_id: req.params.id };
+      con.query(sql, project, (err, result) => {
+        if (err) throw err;
+        res.send(result);
+      });
+    } catch (error) {
+      console.log(error);
+    }
+  } else {
+    res.send("not authorised");
   }
 });
